+++ Assembler/hashtable.h
@@ -7,7 +7,7 @@
 
 struct HashEntry {
     const char *key;
-    uint16_t item;
+    uint32_t item;
 };
 
 #define _HASHSIZE 100
@@ -16,8 +16,8 @@ typedef struct HashEntry *Hashtable;
 Hashtable newHashtable();
 void freeHT(Hashtable ht);
 _Bool htHas(const Hashtable ht, const char *key);
-uint16_t htGet(const Hashtable ht, const char *key);
-void htSet(Hashtable ht, const char *key, uint16_t value);
+uint32_t htGet(const Hashtable ht, const char *key);
+void htSet(Hashtable ht, const char *key, uint32_t value);
 void htPrint(Hashtable ht);
 
 #endif
